requirements.txt
nltk
textblob
tweepy
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-rest-swagger==2.2.0
djangorestframework==3.9.2
psycopg2-binary==2.7.7
requests==2.21.0
django-cors-headers==2.5.2
django-rest-auth==0.9.5
gunicorn==19.9.0
